@@ -41,52 +41,52 @@
 		"react": "16.* || 17.*"
 	},
 	"devDependencies": {
-		"@babel/core": "^7.14.8",
-		"@babel/preset-react": "^7.14.5",
-		"@emotion/react": "^11.4.0",
-		"@emotion/styled": "^11.3.0",
-		"@types/dagre": "^0.7.45",
-		"@types/jest": "^26.0.24",
-		"@types/jest-environment-puppeteer": "^4.4.1",
-		"@types/lodash": "^4.14.171",
-		"@types/node": "^16.4.1",
-		"@types/puppeteer": "^5.4.4",
-		"@types/react": "^17.0.14",
-		"babel-loader": "^8.2.2",
+		"@babel/core": "^7.18.2",
+		"@babel/preset-react": "^7.17.12",
+		"@emotion/react": "^11.9.0",
+		"@emotion/styled": "^11.8.1",
+		"@types/dagre": "^0.7.47",
+		"@types/jest": "^28.1.0",
+		"@types/jest-environment-puppeteer": "^5.0.2",
+		"@types/lodash": "^4.14.182",
+		"@types/node": "^17.0.39",
+		"@types/puppeteer": "^5.4.6",
+		"@types/react": "^18.0.10",
+		"babel-loader": "^8.2.5",
 		"cross-env": "^7.0.3",
-		"css-loader": "^6.2.0",
+		"css-loader": "^6.7.1",
 		"dagre": "^0.8.5",
 		"emotion": "^11.0.0",
 		"enzyme": "^3.11.0",
 		"file-loader": "^6.2.0",
-		"glob": "^7.1.7",
-		"jest": "^27.0.6",
-		"jest-cli": "^27.0.6",
-		"jest-puppeteer": "^5.0",
+		"glob": "^8.0.3",
+		"jest": "^28.1.0",
+		"jest-cli": "^28.1.0",
+		"jest-puppeteer": "^6.1",
 		"json-beautify": "^1.1.1",
-		"lerna": "^4.0.0",
+		"lerna": "^5.0.0",
 		"lodash": "^4.17.21",
 		"pathfinding": "^0.4.18",
 		"paths-js": "^0.4.11",
-		"prettier": "^2.3.2",
-		"puppeteer": "10.1.0",
+		"prettier": "^2.6.2",
+		"puppeteer": "14.2.1",
 		"raf": "^3.4.1",
 		"raw-loader": "^4.0.2",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
-		"react-test-renderer": "^17.0.2",
+		"react": "^18.1.0",
+		"react-dom": "^18.1.0",
+		"react-test-renderer": "^18.1.0",
 		"resize-observer-polyfill": "^1.5.1",
 		"rimraf": "^3.0.2",
-		"source-map-loader": "^3.0.0",
-		"style-loader": "^3.2.1",
-		"terser-webpack-plugin": "^5.1.4",
-		"ts-jest": "^27.0.4",
-		"ts-loader": "^9.2.3",
-		"typescript": "^4.3.5",
-		"val-loader": "^4.0.0",
-		"webpack": "^5.46.0",
-		"webpack-cli": "^4.7.2",
-		"webpack-dev-server": "^3.11.2",
+		"source-map-loader": "^3.0.1",
+		"style-loader": "^3.3.1",
+		"terser-webpack-plugin": "^5.3.3",
+		"ts-jest": "^28.0.4",
+		"ts-loader": "^9.3.0",
+		"typescript": "^4.7.3",
+		"val-loader": "^5.0.0",
+		"webpack": "^5.73.0",
+		"webpack-cli": "^4.9.2",
+		"webpack-dev-server": "^4.9.1",
 		"webpack-node-externals": "^3.0.0"
 	}
 }
