@@ -36,7 +36,7 @@
 	},
 	"peerDependencies": {
 		"lodash": "4.*",
-		"react": "16.* || 17.*",
+		"react": "18.*",
 		"resize-observer-polyfill": "^1.5.1"
 	},
 	"gitHead": "bb878657ba0c2f81764f32901fd96158a0f8352e"
