@@ -36,7 +36,8 @@ export class DiamondNodeWidget extends React.Component<DiamondNodeWidgetProps> {
 					position: 'relative',
 					width: this.props.size,
 					height: this.props.size
-				}}>
+				}}
+			>
 				<svg
 					width={this.props.size}
 					height={this.props.size}
@@ -72,7 +73,8 @@ export class DiamondNodeWidget extends React.Component<DiamondNodeWidgetProps> {
 						position: 'absolute'
 					}}
 					port={this.props.node.getPort(PortModelAlignment.LEFT)}
-					engine={this.props.engine}>
+					engine={this.props.engine}
+				>
 					<S.Port />
 				</PortWidget>
 				<PortWidget
@@ -82,7 +84,8 @@ export class DiamondNodeWidget extends React.Component<DiamondNodeWidgetProps> {
 						position: 'absolute'
 					}}
 					port={this.props.node.getPort(PortModelAlignment.TOP)}
-					engine={this.props.engine}>
+					engine={this.props.engine}
+				>
 					<S.Port />
 				</PortWidget>
 				<PortWidget
@@ -92,7 +95,8 @@ export class DiamondNodeWidget extends React.Component<DiamondNodeWidgetProps> {
 						position: 'absolute'
 					}}
 					port={this.props.node.getPort(PortModelAlignment.RIGHT)}
-					engine={this.props.engine}>
+					engine={this.props.engine}
+				>
 					<S.Port />
 				</PortWidget>
 				<PortWidget
@@ -102,7 +106,8 @@ export class DiamondNodeWidget extends React.Component<DiamondNodeWidgetProps> {
 						position: 'absolute'
 					}}
 					port={this.props.node.getPort(PortModelAlignment.BOTTOM)}
-					engine={this.props.engine}>
+					engine={this.props.engine}
+				>
 					<S.Port />
 				</PortWidget>
 			</div>
