@@ -60,10 +60,12 @@ export default () => {
 				<DemoButton
 					onClick={() => {
 						action('Serialized Graph')(JSON.stringify(model.serialize(), null, 2));
-					}}>
+					}}
+				>
 					Serialize Graph
 				</DemoButton>
-			}>
+			}
+		>
 			<DemoCanvasWidget>
 				<CanvasWidget engine={engine} />
 			</DemoCanvasWidget>
