@@ -68,7 +68,8 @@ export class DefaultNodeWidget extends React.Component<DefaultNodeProps> {
 			<S.Node
 				data-default-node-name={this.props.node.getOptions().name}
 				selected={this.props.node.isSelected()}
-				background={this.props.node.getOptions().color}>
+				background={this.props.node.getOptions().color}
+			>
 				<S.Title>
 					<S.TitleName>{this.props.node.getOptions().name}</S.TitleName>
 				</S.Title>
